[scripts] Split scripts.ts into smaller focused files (#4494)
This PR makes the following changes:

- Splits `packages/scripts/src/helpers/scripts.ts`into smaller files
- `exec.ts` — process execution helpers (`exec`, `fork`, `runTSScript`,
`pgrep`)
- `git.ts` — git helpers (`getCommitHash`, `gitDiff`, `getChangedFiles`)
  - `docker.ts` — all docker helpers
- `package-manager.ts` — package manager helpers (`build`, `setup`,
`packageMngrRun`, `runTestFramework`, `packageMngrPublish`)
- `kubernetes.ts` — `kubectl/kind/helm` install checks and cluster
helpers
  - `helm.ts` — helmfile commands and chart update helpers
  - `certs.ts` — certificate generation and reading helpers
- Updates all internal importers to import directly from the new files
instead of `scripts.ts`
- Deletes `scripts.ts`
- Passes `options.trace` / `options.debug` into `writeTerasliceConfig`
so the Teraslice log level in the generated config matches what was
passed to the test runner

Ref to issue #4383
