+++ packages/scripts/src/index.ts
@@ -1,6 +1,12 @@
 export * from './helpers/packages.js';
 export * from './helpers/misc.js';
-export * from './helpers/scripts.js';
+export * from './helpers/exec.js';
+export * from './helpers/docker.js';
+export * from './helpers/git.js';
+export * from './helpers/package-manager.js';
+export * from './helpers/kubernetes.js';
+export * from './helpers/helm.js';
+export * from './helpers/certs.js';
 export * from './helpers/interfaces.js';
 export * from './helpers/images/index.js';
 export * from './helpers/github.js';
